+++ example/lib/examples/dialogs.dart
@@ -45,33 +45,57 @@ class _DialogsExampleState extends State<DialogsExample> {
               child: const Text("show"),
             ),
             ElevatedButton(
-                onPressed: () {
-                  showDialog(
-                    context: context,
-                    builder: (_) => SkadiConfirmationDialog.danger(
-                      positiveTextStyle: const TextStyle(color: Colors.blue),
-                      title: "Danger action",
-                      confirmText: "Delete",
-                      content: const Text("You want to delete?"),
-                      onConfirm: () async {
-                        var nav = Navigator.of(context);
-                        LoadingOverlayProvider.toggle();
-                        infoLog("you confirm");
-                        await SkadiUtils.wait(1000);
-                        bool success = Random().nextBool();
-                        LoadingOverlayProvider.toggle();
-                        if (success) {
-                          nav.pop();
-                        }
-                      },
-                      onCancel: () {
-                        infoLog("you cancel");
-                        Navigator.pop(context);
-                      },
-                    ),
-                  );
-                },
-                child: const Text("show danger")),
+              onPressed: () {
+                showDialog(
+                  context: context,
+                  builder: (_) => SkadiConfirmationDialog.danger(
+                    positiveTextStyle: const TextStyle(color: Colors.blue),
+                    title: "Danger action",
+                    confirmText: "Delete",
+                    content: const Text("You want to delete?"),
+                    onConfirm: () async {
+                      var nav = Navigator.of(context);
+                      LoadingOverlayProvider.toggle();
+                      infoLog("you confirm");
+                      await SkadiUtils.wait(1000);
+                      bool success = Random().nextBool();
+                      LoadingOverlayProvider.toggle();
+                      if (success) {
+                        nav.pop();
+                      }
+                    },
+                    onCancel: () {
+                      infoLog("you cancel");
+                      Navigator.pop(context);
+                    },
+                  ),
+                );
+              },
+              child: const Text("show danger"),
+            ),
+            ElevatedButton(
+              onPressed: () {
+                showDialog(
+                  context: context,
+                  builder: (_) => SkadiConfirmationDialog(
+                    positiveTextStyle: const TextStyle(color: Colors.blue),
+                    content: const Text("ចុម? លេងអីតាណើបៗចឹង?"),
+                    title: "កាងម៉េស",
+                    confirmText: "អញចឹង",
+                    cancelText: "សូមទោសបង",
+                    onConfirm: () {
+                      infoLog("you confirm");
+                      Navigator.pop(context);
+                    },
+                    onCancel: () {
+                      infoLog("you cancel");
+                      Navigator.pop(context);
+                    },
+                  ),
+                );
+              },
+              child: const Text("ពុម្ពអក្សរខ្មែរ"),
+            ),
           ],
         ),
         Section(