+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
 import 'package:skadi/skadi.dart';
 import 'package:skadi_example/examples/buttons.dart';
 import 'package:skadi_example/examples/controls.dart';
@@ -43,6 +44,7 @@ class MyApp extends StatelessWidget {
           appBarTheme: const AppBarTheme(
             foregroundColor: Colors.white,
           ),
+          fontFamily: GoogleFonts.kantumruy().fontFamily,
         ),
         home: const RootPage(),
         builder: (context, child) {