+++ lib/src/widgets/skadi_async_button.dart
@@ -107,8 +107,7 @@ class _SkadiAsyncButtonState extends State<SkadiAsyncButton> {
     if (widget.fullWidth == false) {
       WidgetsBinding.instance.addPostFrameCallback((d) {
         if (_globalKey.currentContext != null) {
-          RenderBox box =
-              _globalKey.currentContext!.findRenderObject() as RenderBox;
+          RenderBox box = _globalKey.currentContext!.findRenderObject() as RenderBox;
           width = box.size.width;
         }
       });
@@ -128,6 +127,7 @@ class _SkadiAsyncButtonState extends State<SkadiAsyncButton> {
   }
 
   void _toggleLoading(bool value) {
+    if (_isLoading == value) return;
     if (mounted) setState(() => _isLoading = value);
   }
 
@@ -182,9 +182,8 @@ class _SkadiAsyncButtonState extends State<SkadiAsyncButton> {
       ],
     );
 
-    final Widget loadingWidget = widget.loadingWidget ??
-        SkadiProvider.of(context)?.buttonLoadingWidget ??
-        _buildLoadingWidget();
+    final Widget loadingWidget =
+        widget.loadingWidget ?? SkadiProvider.of(context)?.buttonLoadingWidget ?? _buildLoadingWidget();
 
     return Container(
       height: widget.height,
@@ -209,9 +208,7 @@ class _SkadiAsyncButtonState extends State<SkadiAsyncButton> {
         ),
         child: ConditionalWidget(
           condition: _isLoading,
-          onTrue: () => widget.loadingType == LoadingType.disable
-              ? buttonContent
-              : loadingWidget,
+          onTrue: () => widget.loadingType == LoadingType.disable ? buttonContent : loadingWidget,
           onFalse: () => buttonContent,
         ),
       ),